+++ pkg/template/master_template.go
@@ -326,7 +326,11 @@ systemd:
         --container-runtime=remote \
         --container-runtime-endpoint=unix:///run/containerd/containerd.sock \
         --logtostderr=true \
-        --cloud-provider={{.Cluster.Kubernetes.CloudProvider}} \
+        {{- if eq .Cluster.Kubernetes.CloudProvider "aws" }}
+        - --cloud-provider=external
+        {{ else -}}
+        - --cloud-provider={{.Cluster.Kubernetes.CloudProvider}}
+        {{ end -}}
         --pod-infra-container-image={{ .Images.Pause }} \
         --image-pull-progress-deadline={{.ImagePullProgressDeadline}} \
         --network-plugin=cni \