+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Changed
 
+- Use external cloud provider for AWS.
 - Use containerd socket instead of dockershim in the kubelet config.
 - Change liveness probe ports and metric ports of controller-manager and scheduler.
 - Change apiVersion of kubescheduler to `v1beta2`.